Handle timestamp TTLs
Memcached allows expiration times to be specified as either relative to now or as a timestamp.
But since relative times only allows values less than 30 days, the code tried to handle larger valued by converting them to timestamps. This caused actual timestamps to be offset incorrectly.

Now it checks if the values looks like a timestamp (i.e. has a value like a future timestamp) before doing the conversion.