ActiveSupport instrument DalliStore errors
Previously, one could use raise_errors to either handle execptions
in client code or have them handled by dalli. In the later case,
logging was the only way to be notified of an exception. This
commit adds the ability to be notified via active support
instrumentation via a cache_error.active_support event.

I've added this code because I want to know when errors are
occuring to trigger external monitoring and I still want
dalli to fail without raising errors so that cached values
are still generated.