Extracts key validation and namespace handling from the Dalli client. (#820)
Simplifies the Dalli client code a bit
Makes it easier to do an extraction of the multi-yielder, by decoupling the multi-yielder code from the client
Exposed and allowed resolution of issue #819 , which was a bug in the use of alternate digest classes
Makes it simpler to support base64 encoded keys as described in the meta protocol document for a future meta implemenation

To maintain backwards compatibility the key truncation code is more complex than it could be.  Specifically, there is a difference in how clients using namespaces and clients without namespaces truncate keys.  This is essentially due to an off by one error introduced in the historical logic that didn't take into account the namespace separator character.  So without a namespace truncated keys are one character shorter than they need to be (249 characters).  Had the logic been unified and all truncated keys forced to 250 characters in length, that might break some key lookup for some deployed systems, so this complexity was left in place.  This may be revisited in a future major version.