Start extracting connection logic from Dalli::Protocol::Binary (#863)
Once the connection logic is extracted it will become much easier to implement the meta protocol.  Ideally everything that's actually in the Dalli::Protocol::Binary class is related to the binary protocol specifically, and is not reusable for other protocols.  This is basically mapping Ruby commands to memcached binary requests and processing responses, and translating back to appropriate values for the Dalli::Client.

Config parameter parsing, connection handling (initialization, state management, retries), and other shared concerns should be extracted.  Binary protocol specific logic may be extracted into additional classes in the Dalli::Protocol::Binary class scope.

Once this extraction is complete, we may very well wind up inverting this a bit, where we inject the connection manager into a protocol implementation.

This PR also includes some work on the quiet block, adding additional commands that can be used in such a block, as well as making the historical multi an alias for a more onpoint quiet.  There's also a little bit of refactoring of the pipelined gets